# kratos_client/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/cart/create:
        post:
            tags:
                - Cart
            operationId: Cart_CreateCart
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.cart.v1.CreateCartRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.cart.v1.CreateCartReply'
    /v1/cart/delete:
        post:
            tags:
                - Cart
            operationId: Cart_DeleteCart
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.cart.v1.DeleteCartRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.cart.v1.DeleteCartReply'
    /v1/cart/list:
        post:
            tags:
                - Cart
            operationId: Cart_ListCart
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.cart.v1.ListCartRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.cart.v1.ListCartReply'
    /v1/cart/update:
        post:
            tags:
                - Cart
            operationId: Cart_UpdateCart
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.cart.v1.UpdateCartRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.cart.v1.UpdateCartReply'
    /v1/chat/history:
        get:
            tags:
                - Chat
            description: 获取聊天历史记录
            operationId: Chat_GetChatHistory
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: targetId
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/.GetChatHistoryReply'
    /v1/chat/message:
        post:
            tags:
                - Chat
            description: 保存聊天消息
            operationId: Chat_SaveChatMessage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/.SaveChatMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/.SaveChatMessageReply'
    /v1/chat/rooms:
        get:
            tags:
                - Chat
            description: 获取用户的聊天房间列表
            operationId: Chat_GetUserChatRooms
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/.GetUserChatRoomsReply'
...
components:
    schemas:
        .ChatMessageInfo:
            type: object
            properties:
                id:
                    type: string
                fromId:
                    type: integer
                    format: int32
                toId:
                    type: integer
                    format: int32
                fromName:
                    type: string
                toName:
                    type: string
                content:
                    type: string
                messageType:
                    type: string
                roomId:
                    type: string
                createdAt:
                    type: string
        .ChatRoomInfo:
            type: object
            properties:
                roomId:
                    type: string
                targetId:
                    type: integer
                    format: int32
                targetName:
                    type: string
                targetRole:
                    type: string
                lastMessage:
                    type: string
                lastMessageTime:
                    type: string
                unreadCount:
                    type: integer
                    format: int32
...
        .GetChatHistoryReply:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/.ChatMessageInfo'
                total:
                    type: integer
                    format: int32
...
        .GetUserChatRoomsReply:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
                rooms:
                    type: array
                    items:
                        $ref: '#/components/schemas/.ChatRoomInfo'
...
        .SaveChatMessageReply:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
        .SaveChatMessageRequest:
            type: object
            properties:
                fromId:
                    type: integer
                    format: int32
                toId:
                    type: integer
                    format: int32
                content:
                    type: string
                messageType:
                    type: string
                roomId:
                    type: string
        api.cart.v1.CreateCartReply:
            type: object
            properties:
                code:
                    type: string
                msg:
                    type: string
        api.cart.v1.CreateCartRequest:
            type: object
            properties:
                userId:
                    type: string
                drugId:
                    type: string
                number:
                    type: string
        api.cart.v1.DeleteCartReply:
            type: object
            properties:
                code:
                    type: string
                msg:
                    type: string
        api.cart.v1.DeleteCartRequest:
            type: object
            properties:
                id:
                    type: string
        api.cart.v1.InfoCart:
            type: object
            properties:
                id:
                    type: string
                userId:
                    type: string
                drugId:
                    type: string
                number:
                    type: string
        api.cart.v1.ListCartReply:
            type: object
            properties:
                code:
                    type: string
                msg:
                    type: string
                cart:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.cart.v1.InfoCart'
        api.cart.v1.ListCartRequest:
            type: object
            properties:
                userId:
                    type: string
        api.cart.v1.UpdateCartReply:
            type: object
            properties:
                code:
                    type: string
                msg:
                    type: string
        api.cart.v1.UpdateCartRequest:
            type: object
            properties:
                id:
                    type: string
                number:
                    type: string
...